Changed m_signalsChangedInterval back to long int to prevent potential implicit narrowing conversion.
DeviceClient::setDeviceMonitorInterval(long int milliseconds) is a public method and has been like that for a while. Changing its parameter type to int would be a breaking change.

+++ src/karabo/core/DeviceClient.hh
@@ -125,7 +125,7 @@ namespace karabo {
 
             boost::asio::deadline_timer m_signalsChangedTimer;
             bool m_runSignalsChangedTimer;
-            std::atomic<int> m_signalsChangedInterval;
+            std::atomic<long int> m_signalsChangedInterval;
             boost::mutex m_signalsChangedMutex;
             SignalChangedMap m_signalsChanged; /// map of collected signalChanged
